requirements.txt
unittest2
numpy>=1.3.0
scipy>=0.7.0
pyfits>=2.3.1
python-dateutil>=1.4.1
python-monetdb>=11.11.11
pyrap>=1.1.0
sphinx
phinxcontrib-programoutput
pytz
pywcs
pymongo
